separate route for program joining details
